requirements.txt
snowflake-sqlalchemy==1.2.4
apache-airflow-providers-slack==5.1.0
dbt-bigquery==1.5.0
dbt-core==1.5.0